This commit begins the refactoring of the `quanta_tissu/quanta_tissu` directory to `quanta_tissu/tisslm`. It includes the directory rename and updates to the application scripts and some of the test files. This is a partial commit as requested.
