+++ quanta_tissu/scripts/enhanced_demo.py
@@ -5,9 +5,9 @@ import numpy as np
 # Add the project root to the Python path
 sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
 
-from quanta_tissu.quanta_tissu.model import QuantaTissu
-from quanta_tissu.quanta_tissu.tokenizer import Tokenizer, detokenize
-from quanta_tissu.quanta_tissu.config import model_config
+from quanta_tissu.tisslm.model import QuantaTissu
+from quanta_tissu.tisslm.tokenizer import Tokenizer, detokenize
+from quanta_tissu.tisslm.config import model_config
 
 def demonstrate_enhanced_functionality():
     """
