+++ quanta_tissu/scripts/generate_text.py
@@ -6,12 +6,9 @@ import argparse
 # Add the project root to the Python path
 sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
 
-from quanta_tissu.quanta_tissu.model import QuantaTissu
-from quanta_tissu.quanta_tissu.tokenizer import detokenize
-from quanta_tissu.quanta_tissu.config import model_config, vocab
-from quanta_tissu.quanta_tissu.model import QuantaTissu
-
-from quanta_tissu.quanta_tissu.tokenizer import Tokenizer
+from quanta_tissu.tisslm.model import QuantaTissu
+from quanta_tissu.tisslm.tokenizer import detokenize, Tokenizer
+from quanta_tissu.tisslm.config import model_config, vocab
 
 def generate_text(prompt: str, length: int) -> str:
     """
