+++ quanta_tissu/tisslm/knowledge_base.py
